+++ src/Opaleye/Internal/RunQueryExternal.hs
@@ -44,9 +44,9 @@ runQueryFold = runQueryFoldExplicit D.def
 
 queryRunnerColumn :: (Column a' -> Column a) -> (b -> b')
                   -> IRQ.FromField a b -> IRQ.FromField a' b'
-queryRunnerColumn colF haskellF qrc = IRQ.QueryRunnerColumn (P.lmap colF u)
-                                                            (fmapFP haskellF fp)
-  where IRQ.QueryRunnerColumn u fp = qrc
+queryRunnerColumn colF haskellF qrc = IRQ.FromField (P.lmap colF u)
+                                                    (fmapFP haskellF fp)
+  where IRQ.FromField u fp = qrc
         fmapFP = fmap . fmap . fmap
 
 -- * Explicit versions