+++ src/Opaleye/RunSelect.hs
@@ -141,8 +141,8 @@ foldForward = RQ.foldForward
 unsafeFromField :: (b -> b')
                 -> IRQ.FromField sqlType b
                 -> IRQ.FromField sqlType' b'
-unsafeFromField haskellF (IRQ.QueryRunnerColumn u fp) =
-  fmap haskellF (IRQ.QueryRunnerColumn (P.lmap (C.unsafeCoerceColumn) u) fp)
+unsafeFromField haskellF (IRQ.FromField u fp) =
+  fmap haskellF (IRQ.FromField (P.lmap (C.unsafeCoerceColumn) u) fp)
 
 runSelectExplicit :: FromFields fields haskells
                   -> PGS.Connection